requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.6
deepdiff==7.0.1
emoji==2.11.1
exceptiongroup==1.2.1
faiss-cpu==1.8.0
filelock==3.14.0
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain==0.2.0rc2
langchain-community==0.2.0rc1
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.56
lxml==5.2.1
MarkupSafe==2.1.5
marshmallow==3.21.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
ordered-set==4.1.0
orjson==3.10.3
packaging==23.2
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pypdf==4.2.0
pypinyin==0.51.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-iso639==2024.4.27
python-magic==0.4.27
PyYAML==6.0.1
rapidfuzz==3.9.0
regex==2024.5.10
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.40.2
typing-inspect==0.9.0
typing_extensions==4.11.0
unstructured==0.13.7
unstructured-client==0.22.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
zhipuai==2.0.1.20240429
